feat(config): generate JSON schema for the configuration file
This adds a new command line flag that makes presenterm print out a JSON
schema for the configuration.

Usage:

    presenterm --generate-config-file-schema > schema.json

The schema can be used for validation, completion and documentation of
the configuration file.

To use it, you have to have yaml-language-server set up for your editor.
Add a magic comment to the top of the configuration file with the path
to the schema:

    # yaml-language-server: $schema=schema.json